Remove existing alerts from components and add alerts to App and Modal
Closes #56.

We are making this change in large part in order to simplify the code
around alerts: right now alerts involve a fair amount of complexity, as
well as additional communication between components.

Any component that is a modal will still contain an alert, but only
indirectly. For example, the new form modal, FormNew, contains an alert,
but only because it contains a Modal component, which contains the
Alert component.

Note that this commit also removes the login success alert.