Add existing IcoMoon data
I am not sure that this is the best location for this data, but we can
always move it in the future.

Note that the data does not match the IcoMoon assets of this branch: a
commit to follow this one will replace those assets after re-exporting
them from IcoMoon.

Closes #131.