Fix order of test form attachments in FormAttachmentList tests
I think the tests were passing before because the test form attachments'
creation order matched their order as returned by sorted().

After removing get() and using sorted(), the code for an attachment with
the same name and for an attachment with a different name was almost the
same, so I moved that into a function.