Hide modal backdrop after encryption is enabled
When Frontend sends a new request for the project after encryption is
enabled, it sets `project` to `null`. Before, doing so removed the modal
element entirely. However, we only want to hide the modal; hiding the
modal will also hide the backdrop. Thus, we now render the modals for
the Settings tab even if `project` is `null`.