Make changes related to Enketo ID of form draft
I thought that the Enketo ID of a form without a published version was
the same as the Enketo ID of the form draft, but that is not the case.
That means that we do not have access to the Enketo ID of the form draft
when rendering the forms table. Given that, we now hide the preview
button for a form without a published version.