Add ability to switch between Managed and Legacy Code
This commits adds three new components.

The Popover component implements the basic popover functionality and
uses the Bootstrap popover plugin. The Popover component is similar in
some ways to Modal, which implements the basic modal functionality and
uses the Bootstrap modal plugin. The Popover component closes #61. I
also made small changes related to Modal; this commit closes #64 and
closes #118. Note that we no longer implement caching around popovers:
doing so adds complexity, and doesn't seem necessary.

The CollectQr component renders a QR code for Collect. We used to do so
using a function, but now that we can pass HTML to the app user popover
using a slot, CollectQr can be a component. This seems more natural and
also simplifies tests.

The FieldKeyQrPanel component shows the app user QR code. It is shown
within the popover and also within FieldKeyNew.

This commit also adds support for nested component interpolation, which
FieldKeyQrPanel uses.