Add ability to filter submissions
Closes getodk/central#23.

This commit adds new functionality to SubmissionList. However, it also
simplifies SubmissionList in other ways:

- SubmissionAnalyze has been moved to FormSubmissions, and the
  "Analyze via OData" button has been moved to the new component
  SubmissionDataAccess. Relatedly, the request for `keys` has been moved
  to FormSubmissions and FormDraftTesting.
- The message in SubmissionList is now based on computed properties,
  which should make it easier to reason about. This commit closes #212.
- RefreshButton has been replaced with a simpler approach.
- The chunkSizes prop has been replaced with the more general `top`
  prop. It is now easier to track $top and $skip.

The new SubmissionList message load.filtered.middle is pluralized, and
the messages loading.first and loading.middle are now also pluralized.
These messages need to be pluralized for locales that pluralize based on
the number's last digits, such as Arabic and Russian.

This commit adds a .form-label for select elements and for
DateRangePicker. It also adds an optional .close button to
DateRangePicker.

This commit closes #307.

In testing, TestUtilVModel has been replaced with sinon in tests of
DateRangePicker and AuditFiltersAction, and TestUtilVModel has been
removed.