Reduce width of Time column of audit log
Closes #432.

The maximum width of a DateTime component seems to be about 113px,
including in non-English locales. Adding 20% to that just in case, then
adding 16px for the horizontal padding of the td element is about 150px.

Also removing the `details` class from a th element that didn't seem to
use it.