Add container plugin
This is the first step in implementing a container concept. The goal is
to create a container that will make it easier to work with
application-level objects like the router, the store, and the i18n
object. The container will make it easier to pass around those objects,
as well as to mock them in testing. Once the container is fully
implemented, application-level objects will no longer be single, global
objects. For example, these objects will be recreated at the start of
each test rather than reset after each test.