Pass i18n to presenter objects
I initially tried a different approach in which each container defined
its own set of presenter classes, passing `i18n` to the class
definitions. However, that meant that presenter classes could no longer
be imported, so they could no longer be specified as prop types. It also
became more difficult to use presenter classes in testing, because a
test would need to create a container before using a presenter class. As
a result, I ended up simply removing presenter classes from many tests
that didn't absolutely need to use them.

The approach of this commit takes a slightly different approach.
Presenter classes remain untied to a container, so they can continue to
be imported and used as prop types and in tests. If a presenter object
needs to use `i18n`, it can be passed `i18n` through its constructor.
That said, most presenter classes don't use `i18n`, so nothing about the
use of those classes needs to change, especially in tests. When a
presenter object does need to be passed `i18n`, there is a mechanism in
addition to the constructor to facilitate that, bearing resemblance to
the first approach. Each container will define its own set of presenter
subclasses; each subclass will have a static method named from() that
will create a new presenter object and automatically pass in `i18n`.

In e10660f0783a22462f1ce67090bc0831fe4ac7fa, thinking ahead to the first
approach above, I removed the Form presenter from the tests of
EnketoPreview and EnketoFill. Now that we are taking a different
approach, this commit adds the presenter back to those tests.