Account for deletion of i18n properties after component is unmounted
In Vue I18n v9, it looks like i18n properties like $i18n and $t() are
removed from the component after it is unmounted. The only place where
this seems to be an issue for us is alerts: if an alert will be shown
after the component is unmounted, we will need to access the i18n
message beforehand.