Update news, linking to announcement about Google Drive backups (#320)
* Update news, linking to forum announcement about Google Drive backups

* Ensure that news items do not wrap

The news is displayed over two lines, and we want to display two items.
By default, text overflow would cause an item to wrap, so we turn off
wrapping. There isn't text overflow on my machine, but if fonts on
another machine result in text overflow, this will handle it.