Check for vertical truncation in v-tooltip.text
The detectTextOverflow middleware still only modifies x, not y, so the
tooltip won't be positioned correctly if its placement is left or right.
However, we don't need to account for that case yet: in the one place
where vertical truncation is currently possible, the tooltip is
guaranteed to be placed at the top.